IGNITE-23903 Remove setForceServerMode from IgniteClientDataStructuresTest
